Cambio de nombres de las clases, basados en el diagrama
