+++ src/main/java/com/epn/trappi/models/proveedores/CantidadDeBien.java
@@ -1,7 +1,4 @@
 package com.epn.trappi.models.proveedores;
-
-import java.util.ArrayList;
-
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
@@ -12,7 +9,14 @@ import java.util.ArrayList;
  *
  * @author franc
  */
-public class SolicitudCompra {
-        ArrayList <FacturaCompra> facturas = new ArrayList <FacturaCompra>();
+public class CantidadDeBien {
+    private Bien bien;
+    private int cantidad;
 
+    public CantidadDeBien(Bien bien, int cantidad) {
+        this.bien = bien;
+        this.cantidad = cantidad;
+    }
+    
+    
 }
