--- src/main/java/com/epn/trappi/models/proveedores/ListaBienes.java
@@ -1,25 +1,0 @@
-package com.epn.trappi.models.proveedores;
-
-import java.util.ArrayList;
-
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-/**
- *
- * @author franc
- */
-public class ListaBienes {
-
-    private ArrayList<ContadorDeBien> listaDeBienes = new ArrayList<ContadorDeBien>();
-
-    public void añadirBien() {
-
-    }
-
-    public void quitarBien(Bien bien) {
-
-    }
-}
