+++ src/main/java/com/epn/trappi/models/proveedores/ListaCantidadDeBienes.java
@@ -1,15 +1,25 @@
 package com.epn.trappi.models.proveedores;
+
+import java.util.ArrayList;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
 /**
  *
  * @author franc
  */
-public class FacturaCompra {
-    Inventario inventario;
-    ListaBienes listaDeBienes;
+public class ListaCantidadDeBienes {
+
+    private ArrayList<CantidadDeBien> listaCantidadDeBienes = new ArrayList<CantidadDeBien>();
+
+    public void añadirBien() {
+
+    }
+
+    public void quitarBien(Bien bien) {
+
+    }
 }
