+++ src/main/java/com/epn/trappi/models/proveedores/SolicitudDeCompras.java
@@ -1,4 +1,7 @@
 package com.epn.trappi.models.proveedores;
+
+import java.util.ArrayList;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
@@ -9,14 +12,7 @@ package com.epn.trappi.models.proveedores;
  *
  * @author franc
  */
-public class ContadorDeBien {
-    private Bien bien;
-    private int cantidad;
+public class SolicitudDeCompras {
+        ArrayList <Compra> compras = new ArrayList <Compra>();
 
-    public ContadorDeBien(Bien bien, int cantidad) {
-        this.bien = bien;
-        this.cantidad = cantidad;
-    }
-    
-    
 }
